code/requirements.txt
Flask==3.0.0
langchain==0.1.0
pandas==2.1.4
python-dotenv==1.0.0
